refactor(session): replace useEffect with useInterval for session polling
The session fetching logic was moved into a `fetchSession` callback and is now executed periodically using `useInterval` to poll the session every 5 minutes. This improves the reliability of session updates without relying on manual refreshes or component re-renders.
