+++ app/[locale]/_docs/[...path]/error.tsx
