+++ app/[locale]/_docs/[...path]/not-found.tsx
