+++ app/[locale]/_docs/[...path]/style.css
