+++ app/[locale]/_docs/doc-search-bar.tsx
