+++ app/[locale]/_docs/doc-type.ts
