+++ app/[locale]/_docs/error.tsx
