+++ app/[locale]/_docs/theme-swicther.tsx
